adult_income/web_app/requirements.txt
numpy==2.2.4
pandas==2.2.3
joblib==1.4.2
Flask==3.1.0
scikit-learn==1.6.1
mlflow==2.22.0
python-dotenv==1.1.0